requirements.txt
PySide6
pybit>=5.5.0
python-binance>=1.0.17
